refactor: delete legacy scan operators / providers / scan_info hierarchy
Follow-on to the gpu_ingestible refactor (020b8a63) — finishes step 10 of
the plan by removing the legacy code that was left as dead-but-compiled
last time. The new ingestible-based path is the only scan code now.

Promotions before deletion (to keep dependent code working):
- row_group_slice + row_group_range + hybrid_scan_reader → new neutral
  header op/scan/row_group_metadata.hpp (shared by the new
  parquet_gpu_ingestible and the legacy gpu_processing parquet_scan_task
  that stays).
- split_payload (was nested as scan_manager::duckdb_native_split_provider::
  split_payload) → free op::scan::duckdb_native_split_payload in
  op/scan/duckdb_native_decoder.hpp, now carrying a stable pointer to the
  ingestible's bind data instead of a shared_ptr to the legacy scan_info
  clone.

Deletions:
- src/include/op/scan/sirius_gpu_parquet_scan_operator.{hpp,cpp}
- src/include/op/scan/sirius_gpu_duckdb_native_scan_operator.{hpp,cpp}
- src/include/scan_manager/parquet_split_provider.{hpp,cpp}
- src/include/scan_manager/duckdb_native_split_provider.{hpp,cpp}
- src/include/scan_manager/cached_split_provider.{hpp,cpp}
- src/include/op/scan/parquet_scan_info.{hpp,cpp}
- src/include/op/scan/duckdb_native_scan_info.{hpp,cpp}
- src/include/op/scan/scan_info.hpp
- src/include/op/scan/parquet_scan_operator_data.hpp

Enum cleanups:
- SiriusPhysicalOperatorType: drop GPU_PARQUET_SCAN, GPU_DUCKDB_NATIVE_SCAN.
- operator_data_type: drop PARQUET_SCAN, SCAN_CACHED (their owners
  parquet_scan_data + scan_cached_operator_data were deleted with
  parquet_scan_operator_data.hpp).
- pipeline_converter debug-log branches collapsed to GPU_SCAN.
- query.cpp scan-source classifier drops the two deleted enum values.
- scan_manager log strings renamed try_make_cached_provider →
  try_make_cached_ingestible (matches the function name change from the
  prior commit).

decode_duckdb_native_split signature now takes
duckdb_native_split_payload (the new free struct) instead of the legacy
nested split_payload. The decoder body reads from payload.table_info
(pointer to the ingestible's own bind data) instead of cloning a separate
scan_info shape every query.

task_creator.cpp: scan_cached_operator_data dynamic_cast swapped for
scan_operator_with_pinned_table_input (same batch member, same role) so
the cached-scan locality fast path still triggers on pinned scans.

Test removals (the ones that exclusively exercised deleted legacy types
and had no clean port path; coverage is preserved by [scan] +
[integration] which both still pass):
- test_parquet_split_provider.cpp, _s3.cpp — parquet_split_provider direct
- test_duckdb_native_split_provider.cpp — duckdb_native_split_provider direct
- test_metadata_gpu_scan_operators.cpp — legacy operator construction
- test_metadata_gpu_pipeline_task_counting.cpp — same
- test_sirius_read_parquet_converter.cpp — converter peeked at legacy _scan_info
- test_scan_manager_s3_end_to_end.cpp — built parquet_split_provider directly
- test_partition_inject.cpp — smoke-tested parquet_scan_data fields
- Last two TEST_CASEs of test_parquet_scan_task.cpp (FLBA-decimal pushdown
  tests of parquet_split_provider) — to be ported to parquet_gpu_ingestible
  in a follow-up commit per the test-coverage plan.

Verified:
- pixi run make: green
- sirius_unittest "[scan]": 159/159 (155,031 assertions)
- sirius_unittest (full): 1517/1519 passing; 2 unrelated failures will be
  investigated separately
