+++ src/include/op/scan/row_group_metadata.hpp
@@ -1,0 +1,106 @@
+/*
+ * Copyright 2026, Sirius Contributors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#pragma once
+
+// sirius
+#include <io/types.hpp>
+
+// cudf
+#include <cudf/io/datasource.hpp>
+#include <cudf/io/experimental/hybrid_scan.hpp>
+#include <cudf/io/parquet_schema.hpp>
+
+// standard library
+#include <cstddef>
+#include <memory>
+#include <string>
+#include <vector>
+
+namespace sirius::op::scan {
+
+using hybrid_scan_reader = cudf::io::parquet::experimental::hybrid_scan_reader;
+
+//===----------------------------------------------------------------------===//
+// row_group_slice
+//===----------------------------------------------------------------------===//
+/**
+ * @brief Represents a set of row groups within a single parquet file.
+ *
+ * Multiple slices can be bundled together to form a single parquet partition corresponding
+ * to a data batch. Used by @c parquet_gpu_ingestible (new path) and @c parquet_scan_task
+ * (legacy gpu_processing path) — kept in a neutral header so both consumers share without
+ * cross-depending on each other's headers.
+ */
+struct row_group_slice {
+  row_group_slice(std::shared_ptr<cudf::io::parquet::FileMetaData const> file_metadata,
+                  std::string file_path,
+                  std::vector<cudf::size_type> row_group_indices,
+                  std::size_t reserved_uncompressed_bytes,
+                  std::size_t reserved_compressed_bytes,
+                  std::shared_ptr<sirius::io::sirius_ioctx> io_ctx        = nullptr,
+                  std::shared_ptr<sirius::io::sirius_io_object> io_object = nullptr)
+    : file_metadata(file_metadata),
+      file_path(file_path),
+      row_group_indices(std::move(row_group_indices)),
+      reserved_uncompressed_bytes(reserved_uncompressed_bytes),
+      reserved_compressed_bytes(reserved_compressed_bytes),
+      io_ctx(std::move(io_ctx)),
+      io_object(std::move(io_object))
+  {
+  }
+  std::shared_ptr<cudf::io::parquet::FileMetaData const> file_metadata;
+  std::string file_path;
+  std::vector<cudf::size_type> row_group_indices;
+  std::size_t reserved_uncompressed_bytes;
+  std::size_t reserved_compressed_bytes;
+  /// Sirius IO context that minted @c io_object. When non-null, the scan
+  /// operator builds a @c sirius_datasource via @c io_ctx->make_datasource;
+  /// when null, it falls back to @c cudf::io::datasource::create.
+  std::shared_ptr<sirius::io::sirius_ioctx> io_ctx;
+  /// Sirius io_object created by the split provider for this file. Shared
+  /// across every slice from the same parquet file.
+  std::shared_ptr<sirius::io::sirius_io_object> io_object;
+};
+
+//===----------------------------------------------------------------------===//
+// row_group_range
+//===----------------------------------------------------------------------===//
+/**
+ * @brief Represents a set of row groups within a single parquet file.
+ *
+ * Used as the unit of work for the legacy parquet_scan_task path (byte-range preloading).
+ *
+ * @todo This needs to be deleted once Iceberg is integrated into scan manager framework.
+ */
+struct row_group_range {
+  row_group_range(std::size_t file_idx,
+                  std::vector<cudf::size_type> row_group_indices,
+                  std::size_t reserved_uncompressed_bytes,
+                  std::size_t reserved_compressed_bytes)
+    : file_idx(file_idx),
+      row_group_indices(std::move(row_group_indices)),
+      reserved_uncompressed_bytes(reserved_uncompressed_bytes),
+      reserved_compressed_bytes(reserved_compressed_bytes)
+  {
+  }
+  std::size_t file_idx;
+  std::vector<cudf::size_type> row_group_indices;
+  std::size_t reserved_uncompressed_bytes;
+  std::size_t reserved_compressed_bytes;
+};
+
+}  // namespace sirius::op::scan
