+++ src/planner/query.cpp
@@ -46,8 +46,6 @@ void query::build_indices()
       // If it's a scan-like source, add to scan operators vector
       if (source->type == op::SiriusPhysicalOperatorType::DUCKDB_SCAN ||
           source->type == op::SiriusPhysicalOperatorType::ICEBERG_SCAN ||
-          source->type == op::SiriusPhysicalOperatorType::GPU_PARQUET_SCAN ||
-          source->type == op::SiriusPhysicalOperatorType::GPU_DUCKDB_NATIVE_SCAN ||
           source->type == op::SiriusPhysicalOperatorType::GPU_SCAN ||
           source->type == op::SiriusPhysicalOperatorType::CPU_SOURCE) {
         _scan_operators.push_back(source.get());
